+++ src/types/models.ts
@@ -4,10 +4,11 @@ import {
   APIGatewayProxyEvent,
   APIGatewayProxyEventV2,
   Callback,
-  Context,
+  Context as ApiGatewayContext,
 } from "aws-lambda";
 import { APIGatewayProxyResult } from "aws-lambda/trigger/api-gateway-proxy"; // eslint-disable-line import/no-unresolved
 import { AnySchema, Schema, ValidationError } from "joi";
+import { LogLevel } from "ts-injection";
 
 import { BigsbyInstance } from "../bigsby";
 import { ResponseBuilder } from "../response";
@@ -36,7 +37,7 @@ export type ApiEvent = APIGatewayProxyEvent | APIGatewayProxyEventV2;
 
 export type HandlerFunction = (
   event: ApiEvent,
-  context: Context,
+  context: ApiGatewayContext,
   callback?: Callback<ApiResponse>
 ) => Promise<ApiResponse>;
 
@@ -56,9 +57,9 @@ export interface ApiResponse extends Omit<APIGatewayProxyResult, "body"> {
 }
 
 export interface ApiErrorResponseBody {
-  code: number;
-  type: string;
+  code: string;
   message: string;
+  url: string;
 }
 
 export interface ApiHandlerConstructor {
@@ -71,7 +72,7 @@ export type ApiHandlerInvokeFunction = (
 
 export type RawHandlerInvokeFunction = (
   event: ApiEvent,
-  context: Context,
+  context: ApiGatewayContext,
   config: ApiConfig
 ) => Promise<any | ApiResponse>;
 
@@ -79,12 +80,13 @@ export interface ApiHandler {
   invoke: ApiHandlerInvokeFunction;
 }
 
-export interface RequestContext extends Record<string, any> {
+export interface RequestContext {
   rawEvent: APIGatewayProxyEvent | APIGatewayProxyEventV2;
   event: StandardizedEvent;
-  apiGwContext: Context;
+  apiGwContext: ApiGatewayContext;
   config: ApiConfig;
   bigsby: BigsbyInstance;
+  state: Record<string, any>;
 }
 
 export type StandardizedEvent = Omit<ApiEvent, "body"> & {
@@ -143,6 +145,7 @@ export type ApiHookNames = keyof ApiLifecycle;
 
 export interface BigsbyConfig {
   api: ApiConfig;
+  logging: LoggingConfig;
 }
 
 export type HookChain<HookMethod> = HookMethod[];
@@ -152,7 +155,24 @@ export type HookResult = {
   immediate?: boolean;
 } | void;
 
-export type HookInput<InputType> = InputType & { prevResult?: HookResult };
+export type HookInput<InputType> = InputType & {
+  response?: ResponseBuilder;
+};
+
+export interface LoggingConfig {
+  enabled: boolean;
+  level: LogLevel;
+  logger?: BigsbyLogger;
+}
+
+export interface BigsbyLogger {
+  debug: LogFunction;
+  info: LogFunction;
+  warn: LogFunction;
+  error: LogFunction;
+}
+
+export type LogFunction = (msg: string, ...meta: any[]) => void;
 
 export interface ApiConfig {
   request: {
@@ -226,7 +246,7 @@ export interface ApiConfig {
     preResponse?: HookChain<
       (
         inputs: HookInput<{
-          response: ApiResponse;
+          handlerResponse: ApiResponse;
           context: RequestContext;
         }>
       ) => Promise<HookResult>